+++ CHANGELOG.md
@@ -17,12 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * CylindricalSurface3D: line_intersections, linesegment_intersections, plane_intersection
 * Line2D: point_distance
 * Line3D: to_2d
-* ArcEllipse2D: straight_line_area
-* Line3D: skew_to: Verifies if two Line3D are skew
+* Line3D: skew_to (verifies if two Line3D are skew)
 * LineSegment3D: line_interserctions
 * ArcEllipse3D: discretization_points
 * FullArc3D: linesegment_intersections
 * Line: sort_points_along_line
+* ArcEllipse2D: length, point_belongs, abscissa, bounding_rectangle, straight_line_area, discretization_points, reverse
 
 
 ### Fixed
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Step: from stream (use BinaryFile instead of io.BinaryIO)
 * Contour: is_overlapping (consider intersecting_points is empty)
 * LineSegment2D: to_wire (use discretization_points instead of discretise)
+* ArcEllipse2D: to_3d
 * Fix boolean operations when faces are 100% coincident
 * Fix some to_step methods from edges.py and faces.py
 
@@ -70,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * FullArc3D: linesegment_intersections
 * Line2D: sort_points_along_line
 * Line3D: sort_points_along_line
+* ArcEllipse2D: length, point_belongs, abscissa, bounding_rectangle, straight_line_area, discretization_points, reverse
 
 
 ## v0.6.0 [11/7/2022]